requirements.txt
# Web scraping and processing
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.35
newsapi-python==0.2.7
lxml[html_clean]

# NLP and ML
nltk==3.8.1
spacy==3.7.2
transformers==4.51.3 # 4.37.2
datasets==2.16.1
torch==2.2.1
torchtext==0.17.1
sentencepiece==0.1.99
scikit-learn==1.3.2
peft==0.7.1
accelerate==0.25.0
evaluate==0.4.1
rouge-score==0.1.2
bert-score==0.3.13
bitsandbytes==0.41.1
einops==0.7.0
protobuf==4.24.4
tiktoken==0.5.2

# Data processing
pandas==2.1.4
numpy==1.26.3
tqdm==4.66.1
openpyxl==3.1.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1
wandb==0.16.2

# UI
streamlit==1.29.0

# Utilities
python-dotenv==1.0.0

# Development tools
jupyter==1.0.0
pytest==7.4.3 

# LLM
openai==1.78.1